smoke - cleanup some file names
