+++ test/automation/src/code.ts
@@ -7,9 +7,9 @@ import { join } from 'path';
 import * as os from 'os';
 import * as cp from 'child_process';
 import { IDriver, IDisposable, IElement, Thenable, ILocalizedStrings, ILocaleInfo } from './driver';
-import { launch as launchElectron } from './electronDriver';
-import { launch as launchPlaywrightBrowser } from './playwrightBrowserDriver';
-import { launch as launchPlaywrightElectron } from './playwrightElectronDriver';
+import { launch as launchElectron } from './electron';
+import { launch as launchPlaywrightBrowser } from './playwrightBrowser';
+import { launch as launchPlaywrightElectron } from './playwrightElectron';
 import { Logger, measureAndLog } from './logger';
 import { copyExtension } from './extensions';
 import * as treekill from 'tree-kill';
@@ -18,19 +18,19 @@ const rootPath = join(__dirname, '../../..');
 
 export interface LaunchOptions {
 	codePath?: string;
-	workspacePath: string;
+	readonly workspacePath: string;
 	userDataDir: string;
-	extensionsPath: string;
-	logger: Logger;
-	logsPath: string;
-	verbose?: boolean;
-	extraArgs?: string[];
-	remote?: boolean;
-	web?: boolean;
-	legacy?: boolean;
-	tracing?: boolean;
-	headless?: boolean;
-	browser?: 'chromium' | 'webkit' | 'firefox';
+	readonly extensionsPath: string;
+	readonly logger: Logger;
+	readonly logsPath: string;
+	readonly verbose?: boolean;
+	readonly extraArgs?: string[];
+	readonly remote?: boolean;
+	readonly web?: boolean;
+	readonly legacy?: boolean;
+	readonly tracing?: boolean;
+	readonly headless?: boolean;
+	readonly browser?: 'chromium' | 'webkit' | 'firefox';
 }
 
 interface ICodeInstance {
@@ -101,42 +101,6 @@ export async function launch(options: LaunchOptions): Promise<Code> {
 	}
 }
 
-async function poll<T>(
-	fn: () => Thenable<T>,
-	acceptFn: (result: T) => boolean,
-	logger: Logger,
-	timeoutMessage: string,
-	retryCount = 200,
-	retryInterval = 100 // millis
-): Promise<T> {
-	let trial = 1;
-	let lastError: string = '';
-
-	while (true) {
-		if (trial > retryCount) {
-			logger.log('Timeout!');
-			logger.log(lastError);
-			logger.log(`Timeout: ${timeoutMessage} after ${(retryCount * retryInterval) / 1000} seconds.`);
-			throw new Error(`Timeout: ${timeoutMessage} after ${(retryCount * retryInterval) / 1000} seconds.`);
-		}
-
-		let result;
-		try {
-			result = await fn();
-			if (acceptFn(result)) {
-				return result;
-			} else {
-				lastError = 'Did not pass accept function';
-			}
-		} catch (e: any) {
-			lastError = Array.isArray(e.stack) ? e.stack.join(os.EOL) : e.stack;
-		}
-
-		await new Promise(resolve => setTimeout(resolve, retryInterval));
-		trial++;
-	}
-}
-
 export class Code {
 
 	private _activeWindowId: number | undefined = undefined;
@@ -177,7 +141,7 @@ export class Code {
 	}
 
 	async waitForWindowIds(accept: (windowIds: number[]) => boolean): Promise<void> {
-		await poll(() => this.driver.getWindowIds(), accept, this.logger, `get window ids`);
+		await this.poll(() => this.driver.getWindowIds(), accept, `get window ids`);
 	}
 
 	async dispatchKeybinding(keybinding: string): Promise<void> {
@@ -235,10 +199,9 @@ export class Code {
 		const windowId = await this.getActiveWindowId();
 		accept = accept || (result => textContent !== undefined ? textContent === result : !!result);
 
-		return await poll(
+		return await this.poll(
 			() => this.driver.getElements(windowId, selector).then(els => els.length > 0 ? Promise.resolve(els[0].textContent) : Promise.reject(new Error('Element not found for textContent'))),
 			s => accept!(typeof s === 'string' ? s : ''),
-			this.logger,
 			`get text content '${selector}'`,
 			retryCount
 		);
@@ -246,47 +209,47 @@ export class Code {
 
 	async waitAndClick(selector: string, xoffset?: number, yoffset?: number, retryCount: number = 200): Promise<void> {
 		const windowId = await this.getActiveWindowId();
-		await poll(() => this.driver.click(windowId, selector, xoffset, yoffset), () => true, this.logger, `click '${selector}'`, retryCount);
+		await this.poll(() => this.driver.click(windowId, selector, xoffset, yoffset), () => true, `click '${selector}'`, retryCount);
 	}
 
 	async waitForSetValue(selector: string, value: string): Promise<void> {
 		const windowId = await this.getActiveWindowId();
-		await poll(() => this.driver.setValue(windowId, selector, value), () => true, this.logger, `set value '${selector}'`);
+		await this.poll(() => this.driver.setValue(windowId, selector, value), () => true, `set value '${selector}'`);
 	}
 
 	async waitForElements(selector: string, recursive: boolean, accept: (result: IElement[]) => boolean = result => result.length > 0): Promise<IElement[]> {
 		const windowId = await this.getActiveWindowId();
-		return await poll(() => this.driver.getElements(windowId, selector, recursive), accept, this.logger, `get elements '${selector}'`);
+		return await this.poll(() => this.driver.getElements(windowId, selector, recursive), accept, `get elements '${selector}'`);
 	}
 
 	async waitForElement(selector: string, accept: (result: IElement | undefined) => boolean = result => !!result, retryCount: number = 200): Promise<IElement> {
 		const windowId = await this.getActiveWindowId();
-		return await poll<IElement>(() => this.driver.getElements(windowId, selector).then(els => els[0]), accept, this.logger, `get element '${selector}'`, retryCount);
+		return await this.poll<IElement>(() => this.driver.getElements(windowId, selector).then(els => els[0]), accept, `get element '${selector}'`, retryCount);
 	}
 
 	async waitForActiveElement(selector: string, retryCount: number = 200): Promise<void> {
 		const windowId = await this.getActiveWindowId();
-		await poll(() => this.driver.isActiveElement(windowId, selector), r => r, this.logger, `is active element '${selector}'`, retryCount);
+		await this.poll(() => this.driver.isActiveElement(windowId, selector), r => r, `is active element '${selector}'`, retryCount);
 	}
 
 	async waitForTitle(accept: (title: string) => boolean): Promise<void> {
 		const windowId = await this.getActiveWindowId();
-		await poll(() => this.driver.getTitle(windowId), accept, this.logger, `get title`);
+		await this.poll(() => this.driver.getTitle(windowId), accept, `get title`);
 	}
 
 	async waitForTypeInEditor(selector: string, text: string): Promise<void> {
 		const windowId = await this.getActiveWindowId();
-		await poll(() => this.driver.typeInEditor(windowId, selector, text), () => true, this.logger, `type in editor '${selector}'`);
+		await this.poll(() => this.driver.typeInEditor(windowId, selector, text), () => true, `type in editor '${selector}'`);
 	}
 
 	async waitForTerminalBuffer(selector: string, accept: (result: string[]) => boolean): Promise<void> {
 		const windowId = await this.getActiveWindowId();
-		await poll(() => this.driver.getTerminalBuffer(windowId, selector), accept, this.logger, `get terminal buffer '${selector}'`);
+		await this.poll(() => this.driver.getTerminalBuffer(windowId, selector), accept, `get terminal buffer '${selector}'`);
 	}
 
 	async writeInTerminal(selector: string, value: string): Promise<void> {
 		const windowId = await this.getActiveWindowId();
-		await poll(() => this.driver.writeInTerminal(windowId, selector, value), () => true, this.logger, `writeInTerminal '${selector}'`);
+		await this.poll(() => this.driver.writeInTerminal(windowId, selector, value), () => true, `writeInTerminal '${selector}'`);
 	}
 
 	async getLocaleInfo(): Promise<ILocaleInfo> {
@@ -312,6 +275,42 @@ export class Code {
 
 	dispose(): void {
 		this.client.dispose();
+	}
+
+	private async poll<T>(
+		fn: () => Thenable<T>,
+		acceptFn: (result: T) => boolean,
+		timeoutMessage: string,
+		retryCount = 200,
+		retryInterval = 100 // millis
+	): Promise<T> {
+		let trial = 1;
+		let lastError: string = '';
+
+		while (true) {
+			if (trial > retryCount) {
+				this.logger.log('Timeout!');
+				this.logger.log(lastError);
+				this.logger.log(`Timeout: ${timeoutMessage} after ${(retryCount * retryInterval) / 1000} seconds.`);
+
+				throw new Error(`Timeout: ${timeoutMessage} after ${(retryCount * retryInterval) / 1000} seconds.`);
+			}
+
+			let result;
+			try {
+				result = await fn();
+				if (acceptFn(result)) {
+					return result;
+				} else {
+					lastError = 'Did not pass accept function';
+				}
+			} catch (e: any) {
+				lastError = Array.isArray(e.stack) ? e.stack.join(os.EOL) : e.stack;
+			}
+
+			await new Promise(resolve => setTimeout(resolve, retryInterval));
+			trial++;
+		}
 	}
 }
