+++ test/automation/src/index.ts
@@ -26,4 +26,4 @@ export * from './viewlet';
 export * from './localization';
 export * from './workbench';
 export * from './driver';
-export { getDevElectronPath, getBuildElectronPath, getBuildVersion } from './electronDriver';
+export { getDevElectronPath, getBuildElectronPath, getBuildVersion } from './electron';
